Backend/requirements.txt
channels==3.0.2
channels-redis==3.2.0
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.0.1
pytest-asyncio==0.14.0
pytest-django==4.1.0
qrcode